backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
pillow==10.0.0
django-cors-headers==4.0.0
django-storages==1.13.2
boto3==1.28.0
django-redis==5.2.0
gunicorn==20.1.0
drf-spectacular==0.26.5
celery==5.3.1
redis==4.5.4
setuptools<81
python-telegram-bot==20.3